refactor(cmd/generator): modularize histogram generator and integrate telemetrygen
- Split generator.go into focused modules (types, distributions, publisher, etc.)
- Replace manual OTLP with official telemetrygen package integration
- Improve maintainability with single-responsibility design
- Add comprehensive examples and documentation
